Extracted PaymentModule to seperate file
