+++ Frontend/frontend/src/Components/Payments.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import PaymentModule from "./PaymentModule";
 import NewPayment from "./NewPayment";
 import ViewPayment from "./ViewPayment";
 import "./Payments.css";
@@ -163,72 +164,6 @@ export default class Payments extends Component {
             </div>
         );
     }
-}
-
-/*
-    This function JSX element takes in the input:
-    - Payment: The list obtained from the backend(GET api/Payments/User)
-    - UserPayment: List of usernames as contributors.
-*/
-function PaymentModule(props) {
-    const onTableClick = props.onTableClick
-
-    // Map enums stored in backend to list.
-
-    const payment = props.payment
-    const paymentType = paymentTypeEnumList[props.payment["paymentType"]];
-    const amount = props.payment["amount"];
-    const startDate = props.payment.startDate
-        .slice(0, 10)
-        .split("-")
-        .join("/");
-    const endDate = props.payment.endDate
-        .slice(0, 10)
-        .split("-")
-        .join("/");
-    const frequency = frequencyEnumList[props.payment["frequency"]];
-
-    //Ensure that Payments Page is rendered even when the list does not exist.
-    const contributors = !props.contributors
-        ? ["Loading..."]
-        : props.contributors;
-    const contributorsToString = contributors.join(", ");
-
-    return (
-        <div className="PaymentModule">
-            <table className="PaymentModule" onClick={() => onTableClick(payment, contributors)}>
-                <tr>
-                    <td className="PaymentModuleDataLeft">
-                        <h6 className="PaymentModuleHeader">
-                            <b>{paymentType}</b>
-                        </h6>
-                    </td>
-                    <td className="PaymentModuleDataRight">
-                        <h6 className="PaymentModuleHeader">
-                            <b>
-                                {startDate}-{endDate}
-                            </b>
-                        </h6>
-                    </td>
-                </tr>
-                <tr>
-                    <td className="PaymentModuleDataLeft">
-                        <h6 className="PaymentModuleData">Amount: ${amount}</h6>
-                    </td>
-                    <td className="PaymentModuleDataRight">
-                        <h6 className="PaymentModuleData">Frequency: {frequency}</h6>
-                    </td>
-                </tr>
-                <tr>
-                    <td colSpan="2" className="PaymentModuleDataLeft">
-                        <h6 className="PaymentModuleData">
-                            Contributors: {contributorsToString}
-                        </h6>
-                    </td>
-                </tr>
-            </table>
-        </div>
-    );
 }
 
 export const paymentTypeEnumList = [
